Clean up clang-tidy warnings from 1D/

+++ 1D/Population.cpp
@@ -23,10 +23,10 @@ double NewNu (vector<double> & nu, double K, double p, int size) {
 	}
 
     double u = prng.uniform_real()*(BB+DD+MM);
-    if (u < BB) {        	int z=0; double v = prng.uniform_real()*BB; while (v>0) { z++; v -= B[z]; } nu[z] += 1/K; }
-    else if (u < BB+DD) {	int z=0; double v = prng.uniform_real()*DD; while (v>0) { z++; v -= De[z]; } nu[z] -= 1/K; if (nu[z]<.5/K) nu[z]=0; }
-    else {               	int z=0; double v = prng.uniform_real()*MM; while (v>0) { z++; v -= M[z]; }
-                         	int dz = 2*prng.bernoulli()-1; if ((z>0) && (z+dz>=1) && (z+dz<=size)) nu[z+dz] += 1/K; }
+    if (u < BB) {        	long z=0; double v = prng.uniform_real()*BB; while (v>0) { z++; v -= B[z]; } nu[z] += 1/K; }
+    else if (u < BB+DD) {	long z=0; double v = prng.uniform_real()*DD; while (v>0) { z++; v -= De[z]; } nu[z] -= 1/K; if (nu[z]<.5/K) nu[z]=0; }
+    else {               	long z=0; double v = prng.uniform_real()*MM; while (v>0) { z++; v -= M[z]; }
+                         	long dz = 2*prng.bernoulli()-1; if ((z>0) && (z+dz>=1) && (z+dz<=size)) nu[z+dz] += 1/K; }
 
     return prng.exponential (K*(BB+DD+MM));
 }
@@ -41,7 +41,7 @@ int main (int argc, char ** argv) {
 	vector<double> nu (size+1,0); nu[shift] = 3;
 
 	vector<cpx> pos (size,0);
-	for (int i=0; i<size; ++i) pos[i] = cpx(i,nu[i+1]);
+	for (long i=0; i<size; ++i) pos[i] = cpx(i,nu[i+1]);
 
 	Console W;
 	W.watch (slope, "Slope"); W.manage (slope,-100.0,100.0);
@@ -63,13 +63,13 @@ int main (int argc, char ** argv) {
 		if (!(k%(iterations/1000))) {
 			cout << k << " " << t; for (int i=1; i<=size; ++i) cout << " " << int(.2+nu[i]*K); cout << endl;
 			G.contents.clear();
-			for (int i=0; i<size; ++i) {
+			for (long i=0; i<size; ++i) {
 				graph[i].emplace_back (1000.0*k/iterations,nu[i+1]);
 				G.add (std::make_unique <Path> (graph[i],Pen(Indexed(i))));
 			}
 			if (! G.visible()) { G.show(); } G.step();
 		}
-		for (int i=0; i<size; ++i) pos[i] = cpx(i,nu[i+1]);
+		for (long i=0; i<size; ++i) pos[i] = cpx(i,nu[i+1]);
 		F.contents.clear(); F.add (std::make_unique <Path> (pos));
 		F.step();
 	}
